% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sporulation.R
\name{predict_sporulation}
\alias{predict_sporulation}
\title{Predict Sporulation Potential}
\usage{
predict_sporulation(binary_matrix)
}
\arguments{
\item{binary_matrix}{A binary matrix (1/0) indicating gene presence/absence for each MAG. Must include a \code{genome_ID} column.}
}
\value{
A tibble with predicted class and probability of sporulation for each genome.
}
\description{
This function predicts the sporulation potential of MAGs using an ensemble learning model.
It uses probabilities from a Random Forest and SVM as inputs to a meta-model.
}
