% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sporulation_gene_name.R
\name{sporulation_gene_name}
\alias{sporulation_gene_name}
\title{Identify Sporulation-Associated Genes}
\usage{
sporulation_gene_name(df)
}
\arguments{
\item{df}{A data frame containing MAG annotation with the columns 'Preferred_name' and 'KEGG_ko'.}
}
\value{
A data frame of sporulation-associated genes with standardized names and spo_process tags.
}
\description{
This function identifies sporulation-associated genes in a genome annotation data frame.
It searches for gene names and KEGG Orthology identifiers related to sporulation steps and returns
a data frame with annotated sporulation genes and a consensus name.
}
